Change float to decimal and added unit tests

+++ GustavoSeguros.Application/UseCases/CalcularSeguro/CalcularSeguroPresenter.cs
@@ -7,14 +7,14 @@ namespace GustavoSeguros.Application.UseCases.CalcularSeguro
 {
   public class CalcularSeguroPresenter : IOutputPort
   {
-    public float Valor { get; private set; }
+    public decimal Valor { get; private set; }
     public bool InvalidOutput { get; private set; }
     public void Invalid()
     {
       this.InvalidOutput = true;
     }
 
-    public void Ok(float valor)
+    public void Ok(decimal valor)
     {
       Valor = valor;
     }
