+++ GustavoSeguros.Domain/Veiculo.cs
@@ -5,14 +5,14 @@ namespace GustavoSeguros.Domain
   public class Veiculo
   {
     public int Id { get; set; }
-    public float Valor { get; set; }
+    public decimal Valor { get; set; }
     public string Modelo { get; set; }
 
     public Veiculo()
     {
     }
 
-    public Veiculo(int id, float valor, string modelo)
+    public Veiculo(int id, decimal valor, string modelo)
     {
       Id = id;
       Valor = valor;
